Builds a WFS request to the stat.fi geospatial API.

Builds a WFS request to the stat.fi geospatial API. The request is submitted with GeoStatFiWFSClient to retrieve the data. For more information about the API, see http://www.stat.fi/tup/rajapintapalvelut/index_en.html. For General Terms of Use, see http://www.stat.fi/org/lainsaadanto/yleiset_kayttoehdot_en.html/.

GeoStatFiWFSRequest

Format

An object of class R6ClassGenerator of length 24.

Currently supported data sets

Population densities in various demographic groups

Production and industrial facilities

Educational institutions

Road accidents

Methods

getPopulationLayers(): A request for a list of available population grid data sets (layers).

getPopulation(layer): A request for population grid data layer.

getProductionAndIndustrialFacilities(): A request for production and industrial facilities.

getEducationalInstitutions(): A request for educational institutions.

getRoadAccidentsLayers(): A request for a list of available road accident data sets as a character vector.

getRoadAccidents(layer): A request for road accident data layer.

getPostalCodeAreaLayers(): A request for a list of available postal code area data sets as a character vector.

getPostalCodeArea(layer): A request for postal code area data layer.

References

See citation("gisfin")

See also

Examples

  # See the vignette.