Get geocode for given query

Get gecode for given query from one of the geocoding services: OKF.fi Geocoding API Test Console: http://data.okf.fi/console/ OpenStreetMap Nominatim (usage policy: http://wiki.openstreetmap.org/wiki/Nominatim_usage_policy) Google Maps API See Terms and Conditions from http://code.google.com/apis/maps/documentation/geocoding/

get_geocode(query, service = "okf", raw_query = F)

Arguments

query
Either a street address, e.g. 'Mannerheimintie 100, Helsinki' or place, e.g. 'Eduskuntatalo'

service
Geocoding service to use, one of 'okf', 'openstreetmap' or 'google'

raw_query
If true, don't prepend / append some default parameters to query. Except for the ones specifying json format, send the query to API as-it-is

Value

A list with coordinates (lat, long) of the first output, and the raw output list

References

See citation("gisfin")

Examples

 
# NOT RUN {
   # Geocode for a street address:
   gc1 <- get_geocode("Mannerheimintie 100, Helsinki");

   # Geocode for a place name
   gc2 <- get_geocode("&city=Helsinki", service="openstreetmap", raw_query=T)
# }