Retrieve postal code areas as spatial data

Retrieves postal code areas from http://www.palomaki.info/apps/pnro/ The data is created by Duukkis and licensed under CC BY 4.0 Please not that this function requires a recent version of GDAL. Please see the tutorial for more instructions

get_postalcode_areas(data.dir = tempdir(), verbose = TRUE)

Arguments

data.dir
A string. Specify a temporary folder for storing downloaded data.

verbose
logical. Should R report extra information on progress?

Value

a spatial object (from SpatialPolygonsDataFrame class)

References

See citation("gisfin")

Examples

pnro.sp <- get_postalcode_areas()

#> Finnish postal code areas by Duukkis from http://www.palomaki.info/apps/pnro/

#> Version 2.1.2015

#> License: CC BY 4.0

#> Downloading https://fusiontables.google.com/exporttable?query=select+col2+from+1m9trydaf6rdt9qBfQkU-4iRKkhNn6O1lRw7XWnXF&o=kml&g=col2&styleId=2&templateId=2
#> into /tmp/RtmpPdjRNL/pnro.kml

#> OGR data source with driver: LIBKML 
#> Source: "/tmp/RtmpPdjRNL/pnro.kml", layer: "Postinumerot 20150102"
#> with 3039 features
#> It has 12 fields

#> Warning: Z-dimension discarded

#> 
#> Data loaded successfully!